cmd, ethdb, replica: Port replica changes to Geth 1.9.0
The replica2 branch was based on geth 1.8.27. Initially I tried
to merge 1.9.0 into that branch, but 1.8.27 and 1.9.0 diverged
long enough ago that this approach seemed infeasible. Instead,
I've manually copied the difference between replica2 and 1.8.27
on top of the 1.9.0 branch. Note that this commit almost certainly
won't compile - it doesn't address any interface changes between
1.8.27 and 1.9.0 - it's just a straight copy of the changes, and
I wanted to git a check point before I start trying to make it
compatible with 1.9.0.

Since we aren't retaining the history of the replica2 branch, a brief
description of the branch follows:

The replica branch creates a streaming replica process through Kafka.
Masters can specify a kafka broker and topic, and all database write
operations will be written to Kafka in addition to the local database.

Separately, we have replicas that pull from Kafka and write to local
disks. These replicas have a variation of the ethapi.Backend interface
that pulls everything it needs from disk, and does not rely on p2p
connectivity for anything.

The result is that the master handles p2p connectivity, block validation,
state management, etc. and simply streams the disk writes to replicas.
The replicas are able to scale up relatively quickly to handle RPC
workloads.

Additionally, there is a txrelay command which pulls transactions from
a Kafka topic and transmits them to a Geth node. Transactions from replicas
are placed on this topic to be broadcast to the rest of the network.