ethdb/cdc: Limit overwrites of LastHeader and LastBlock
We want to be able to run multiple masters to support the same set
of replicas. Pretty much every write to leveldb can be repeated, with
the exceptio nof LastHeader and LastBlock, which we don't want to
happen out of order.

This will track recently written values of LastBlock and LastHeader,
and if the value coming from Kafka has been written within our memory
we won't write it again. Thus, the "latest" block should always be
the most recent block from any of the masters contributing to this
replica.