ethdb/overlay, replica: Fix freezer integration
We discovered that when we restarted a replica that had been running
with an overlay, we would get some problems pairing the underlay with
the freezer. The freezer would check that the underlay was in sync
with the freezer, which it wouldn't be because the freezer was updated
in conjunction with the overlay, not the underlay. The freezer would
try to repair the underlay, and would fail.

This changes us from a model of wrapper(overlay, underlay, freezer(underlay))
to freezer(wrapper(overlay, underlay)). The freezer now tries to check
for consistency across the wrapper of the overlay and underlay, rather
than just looking at the underlay where there will be inconsistencies.