replica, internal/ethapi: Tie eth_call execution time to specified gas
We've seen cases where eth_calls take longer to execute than the 5s
timeout. To allow users to bump their execution time, we've tied it
to requested gas. If gas is specified, each 10,000,000 gas gets one
second of execution time, with a minimum of 5s. If gas is not specified,
the transaction gets 5s of execution time.