internal/ethapi: Speculative transaction execution
Since replicas don't run with the miner system, "pending" blocks
are identical to the "latest" block. But for the purposes of
processing eth_calls, some users want the behavior of having
the pending block include the transactions they submitted.

Now, your local node will generally include all of the "local"
transactions in its "pending" blocks. At the scale of Rivet and
with replication, including all of the transactions submitted
directly to us may not be feasible. Instead, if the caller submits
a "from" argument to the eth_call as a pending block is specified,
we will include any pending / queued transactions from that
submitter before running the call.