Backfill estimateGasList into 1.9.12
MEW's estimateGasLIst was built on top of 1.9.14, and given some of
the eth_call refactors that have happened since 1.9.12 we couldn't
merge easily, so I backported it by hand.

Note that this puts it under ethercattle_estimateGasList, so that if
the Geth team ever merges MEW's PR there's no risk of conflicting APIs.