internal/ethapi: Update intermediate state roots
This step was in MEW's original PR. I dropped it when trying to figure
out the bug about state not being retained between steps when the last
iteration of the binary search resulted in out-of-gas. The removal of
this line wasn't important to the ultimate solution, but it got committed
nonetheless.

I'm not entirely sure why this changes the gas estimation - it seems to
be the difference between the EVM treating subsequent transactions as
continuations of the previous ones and treating them as new transactions,
which results in slightly higher gas usage (typically < 10k gas difference
on large transactions).