internal/ethapi: Optimize gas estimation
Experimenting here, but I believe we can use result.UseGased as a
lower bound for the gas estimation (I don't think there's ever a case
where the gas required can be less than the gas used).

Additionally, for estimateGasList, we're going to get with 1% of the
correct value and return the high - That's good enough for most use
cases, and saves several iterations of the estimator.