replica: Pull from successes in the same select as sending inputs
There was a deadlock where the events producer was stalled sending
to the Inputs() channel while the sarama goroutine was stalled sending
to the Successes() channel. Both need to happen at the same time, but
if we take off the successes channel we need to loop back around
to send the same message on the Inputs channel.