Enable creating snapshots in a separate database
Snapshots as currently implemented in Geth don't work well with our
streaming replication system. Only older data is written to the snapshot
database, which means it's not useful on the replicas. We'd still like
to have it on the masters though for performance reasons, so we
want to store it somewhere that it won't be streamed to replicas.